dmaengine: sun6i: Add support for the D1 variant
So far it appears to match the configuration of the A100 variant.

Since D1 is a RISC-V chip, it does not meet any of the existing
dependencies for this driver, so relax the dependency somewhat.

Cover-letter:
dmaengine: sun6i: Allwinner D1 support
D1 is a new RISC-V SoC that uses mostly the same peripherals as
existing ARM-based sunxi SoCs. This series adds dmaengine support for
D1, after fixing an issue where the driver depended on architecture-
specific behavior (patch 2) and resolving a TODO item (patch 3).
END

Series-to: Vinod Koul <vkoul@kernel.org>
Series-to: dmaengine@vger.kernel.org

Series-version: 2

Acked-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Samuel Holland <samuel@sholland.org>