clk: sunxi-ng: Allow the CCU core to be built as a module
Like the individual CCU drivers, it can be beneficial for memory
consumption of cross-platform configurations to only load the CCU core
on the relevant platform. For example, a generic arm64 kernel sees the
following improvement when building the CCU core and drivers as modules:

  before:
    text      data     bss     dec       hex      filename
    13882360  5251670  360800  19494830  12977ae  vmlinux

  after:
    text      data     bss     dec       hex      filename
    13734787  5086442  360800  19182029  124b1cd  vmlinux

So the result is a 390KB total reduction in kernel image size.

The one early clock provider (sun5i) requires the core to be built in.

Now that loading the MMC driver will trigger loading the CCU core, the
MMC timing mode functions do not need a compile-time fallback.

Series-version: 3
Cover-letter:
clk: sunxi-ng: Module support
This series allows the CCU core and drivers to be loaded/unloaded as
modules. As part of this, patch 3 converts most of the early OF clock
providers to platform drivers.
END

Series-to: linux-sunxi
Series-cc: Michael Turquette <mturquette@baylibre.com>
Series-cc: Stephen Boyd <sboyd@kernel.org>
Series-cc: linux-clk@vger.kernel.org
Series-cc: lakml, lkml

Signed-off-by: Samuel Holland <samuel@sholland.org>