phy: rockchip-inno-usb2: Handle ID IRQ
This supports detecting host mode for the OTG port without an extcon.

The rv1108 properties are not updated due to lack of documentation.

Cover-letter:
phy: rockchip-inno-usb2: RK356x OTG fix and enhancements
This series gets USB OTG working on the RK356x SoCs, as used in the
PineNote tablet.

 - Patch 1 fixes an interrupt storm issue specific to RK356x.
 - Patches 2 and 3 are a couple of small optimizations.
 - Patches 4 and 5 reduce the delay when unplugging from the OTG port.
 - Patch 6 allows OTG ports to work in host mode without depending on
   another driver to provide an extcon. The specific use case here is
   a Type-C port controller that provides a hardware "ID pin" output.
END

Series-to: Kishon Vijay Abraham I <kishon@ti.com>
Series-to: Vinod Koul <vkoul@kernel.org>
Series-to: Heiko Stuebner <heiko@sntech.de>

Signed-off-by: Samuel Holland <samuel@sholland.org>