iio: accel: st_accel: Add support for Silan SC7A20
This chip appears to be register-compatible with the LIS2DH. The new
description is a copy of the LIS2DH's description with a different WAI
value.

Series-to: Jonathan Cameron <jic23@kernel.org>
Series-to: Lars-Peter Clausen <lars@metafoo.de>
Series-to: Denis Ciocca <denis.ciocca@st.com>
Series-to: Linus Walleij <linus.walleij@linaro.org>
Series-to: devicetree
Series-cc: linux-iio@vger.kernel.org
Series-cc: lkml

Cover-letter:
iio: Silan SC7A20 accelerometer support
This chip is register-compatible with an ST part, so it works without
any issues when added to the existing driver.

Previously I called it a clone, but it's not really a clone, because
it's not a drop-in replacement for the other chips. The software
interface matches the ST parts, but the pinout does not. Instead, the
pinout matches the Kionix KX122 and KX132.
END

Series-version: 2
Series-changes: 2
 - Add a comment about the WAI value
 - Update commit messages

Datasheet: http://linux-chenxing.org/silan/SC7A20-SilanMicroelectronics.pdf
Datasheet: http://www.siitek.com.cn/Upfiles/down/SC7A20说明书_0.92(智能穿戴).pdf
Signed-off-by: Samuel Holland <samuel@sholland.org>