riscv: Fix irq_work when SMP is disabled
irq_work is triggered via an IPI, but the IPI infrastructure is not
included in uniprocessor kernels. As a result, irq_work never runs.
Fall back to the tick-based irq_work implementation on uniprocessor
configurations.

Series-to: Palmer Dabbelt <palmer@dabbelt.com>
Series-to: linux-riscv@lists.infradead.org

Commit-notes:
This was found while bringing up cpufreq on D1. Switching cpufreq
governors was hanging on irq_work_sync().
END

Fixes: 298447928bb1 ("riscv: Support irq_work via self IPIs")
Signed-off-by: Samuel Holland <samuel@sholland.org>