phy: allwinner: phy-sun6i-mipi-dphy: Add the A100 DPHY variant
A100 features an updated DPHY, which moves PLL control inside the DPHY
register space (previously the PLL was controlled from the CCU). It also
requires a modified analog power-on sequence. This "combo PHY" can also
be used as an LVDS PHY, but that is not yet supported by the driver.

Cover-letter:
phy: allwinner: phy-sun6i-mipi-dphy: Add the A100 DPHY
This series adds support for the updated DPHY found in a couple of
recent Allwinner SoCs. The first three patches fix an omission in the
existing binding. The remaining patches add the new hardware variant.
END

Series-version: 2

Series-to: Kishon Vijay Abraham I <kishon@kernel.org>
Series-to: Vinod Koul <vkoul@kernel.org>
Series-to: Chen-Yu Tsai <wens@csie.org>
Series-to: Jernej Skrabec <jernej.skrabec@gmail.com>
Series-to: Maxime Ripard <mripard@kernel.org>
Series-cc: Paul Kocialkowski <paul.kocialkowski@bootlin.com>

Signed-off-by: Samuel Holland <samuel@sholland.org>