riscv: dts: allwinner: d1: Add crypto engine node
D1 contains a crypto engine which is supported by the sun8i-ce driver.

Series-version: 2

Commit-changes: 2
 - New patch for v2

Series-to: Corentin Labbe <clabbe.montjoie@gmail.com>
Series-to: Herbert Xu <herbert@gondor.apana.org.au>
Series-to: David S. Miller <davem@davemloft.net>
Series-to: Chen-Yu Tsai <wens@csie.org>
Series-to: Jernej Skrabec <jernej.skrabec@gmail.com>
Series-to: Krzysztof Kozlowski <krzysztof.kozlowski+dt@linaro.org>
Series-to: Rob Herring <robh+dt@kernel.org>

Cover-letter:
crypto: Allwinner D1 crypto support
This series finishes adding crypto support for the Allwinner D1 SoC. The
driver patch from v1 was merged, but not the binding[1]. This turned out
to be a good thing, because I later found that the TRNG needed another
clock reference in the devicetree. That is fixed in v2. I include the DT
update here too, since the SoC DT has been on the list for a while[2].

[1]: https://lore.kernel.org/all/20211119051026.13049-1-samuel@sholland.org/T/
[2]: https://lore.kernel.org/lkml/20221208090237.20572-1-samuel@sholland.org/
END

Signed-off-by: Samuel Holland <samuel@sholland.org>