riscv: dts: allwinner: d1: Add RGB LEDs to boards
Some D1-based boards feature an onboard RGB LED. Enable them.

Commit-changes: 5
 - New patch for v5

Cover-changes: 6
 - Drop the A100 DMA controller DT node patch, which was merged via a
   different series

Cover-changes: 8
 - Collected tags, except Tested-by for the driver due to large changes

Series-version: 8

Series-to: Lee Jones <lee@kernel.org>
Series-to: Pavel Machek <pavel@ucw.cz>
Series-to: linux-leds@vger.kernel.org
Series-to: Chen-Yu Tsai <wens@csie.org>
Series-to: Jernej Skrabec <jernej.skrabec@gmail.com>

Cover-letter:
leds: Allwinner A100 LED controller support
This series adds bindings and a driver for the RGB LED controller found
in some Allwinner SoCs, starting with A100. The hardware in the R329 and
D1 SoCs appears to be identical.

This series was tested on the Nezha board (one onboard LED) and the
Dongshan Nezha STU board (with an external strip of 16/32/48 LEDs, also
with gaps in the LED addresses).
END

Acked-by: Guo Ren <guoren@kernel.org>
Acked-by: Jernej Skrabec <jernej.skrabec@gmail.com>
Tested-by: Trevor Woerner <twoerner@gmail.com>
Signed-off-by: Samuel Holland <samuel@sholland.org>