dt-bindings: sun6i-a31-mipi-dphy: Add the interrupts property
The sun6i DPHY can generate several interrupts, mostly for reporting
error conditions, but also for detecting BTA and UPLS sequences.
Document this capability in order to accurately describe the hardware.

The DPHY has no interrupt number provided in the vendor documentation
because its interrupt line is shared with the DSI controller.

Signed-off-by: Samuel Holland <samuel@sholland.org>