arm64: dts: allwinner: a100: Add I2C DMA requests
The I2C controllers in the A100 SoC are all connected to the DMA engine.

Cover-letter:
DMA binding for Allwinner V536/newer I2C
This series documents the DMA capability of the I2C offload engine in
Allwinner V536 and newer SoCs. This includes the following SoCs which
have some support in mainline Linux:
 - A100 (updated by this series)
 - H616 (missing DMA controller binding/support)
 - D1 (initial DT submission is on the ML; this series resolves some
   dtbs_check warnings with that DT)
END

Series-to: Chen-Yu Tsai <wens@csie.org>
Series-to: Jernej Skrabec <jernej.skrabec@gmail.com>

Signed-off-by: Samuel Holland <samuel@sholland.org>