dt-bindings: net: sun8i-emac: Add phy-supply property
This property has always been supported by the Linux driver; see
commit 9f93ac8d4085 ("net-next: stmmac: Add dwmac-sun8i"). In fact, the
original driver submission includes the phy-supply code but no mention
of it in the binding, so the omission appears to be accidental. In
addition, the property is documented in the binding for the previous
hardware generation, allwinner,sun7i-a20-gmac.

Document phy-supply in the binding to fix devicetree validation for the
25+ boards that already use this property.

Series-to: Chen-Yu Tsai <wens@csie.org>
Series-to: Jernej Skrabec <jernej.skrabec@gmail.com>
Series-to: David S. Miller <davem@davemloft.net>
Series-to: Eric Dumazet <edumazet@google.com>
Series-to: Jakub Kicinski <kuba@kernel.org>
Series-to: Paolo Abeni <pabeni@redhat.com>
Series-to: Rob Herring <robh+dt@kernel.org>
Series-to: Krzysztof Kozlowski <krzysztof.kozlowski+dt@linaro.org>

Cover-letter:
dt-bindings: net: sunxi: Fix binding validation issues
These patches fix issues found when running `make dtbs_check` on sunxi
boards. I am not 100% sure about the change in the first two patches,
but I think I understand what is happening, and the change has the
desired effect. It fixes errors of the form:

  arch/arm/boot/dts/sun6i-a31s-colorfly-e708-q1.dtb: ethernet@1c30000:
  Unevaluated properties are not allowed ('mdio', 'reset-names',
  'resets', 'snps,fixed-burst', 'snps,force_sf_dma_mode', 'snps,pbl'
  were unexpected)

where all of the listed properties are defined in snps,dwmac.yaml.

The third patch fixes an unrelated omission in the sun8i-emac binding.
END

Fixes: 0441bde003be ("dt-bindings: net-next: Add DT bindings documentation for Allwinner dwmac-sun8i")
Signed-off-by: Samuel Holland <samuel@sholland.org>