riscv: dts: allwinner: d1: Add video engine node
D1 contains a video engine which is supported by the Cedrus driver.

Series-to: Chen-Yu Tsai <wens@csie.org>
Series-to: Jernej Skrabec <jernej.skrabec@gmail.com>
Series-to: Paul Kocialkowski <paul.kocialkowski@bootlin.com>
Series-to: Mauro Carvalho Chehab <mchehab@kernel.org>
Series-to: Rob Herring <robh+dt@kernel.org>
Series-to: Krzysztof Kozlowski <krzysztof.kozlowski+dt@linaro.org>

Cover-letter:
Allwinner D1 video engine support
This series finishes adding Cedrus support for Allwinner D1. I had
tested the hardware and documented the compatible string a while back,
but at the time I had a dummy SRAM section in the devicetree. Further
testing shows that there is no switchable SRAM section -- there is no
need for it, I was unable to guess the address, and the usual bits in
the SRAM controller register have no effect on the video engine. So that
needs to be made optional in the binding and driver.

With that done, the node can be added to the devicetree.
END

Signed-off-by: Samuel Holland <samuel@sholland.org>