firmware/psci: Switch to the sys-off handler API
Any other poweroff handlers registered at the default priority will
run before the legacy pm_power_off() function. Register the PSCI
poweroff handler with the correct priority to ensure it runs first.

PSCI_0_2_FN_SYSTEM_OFF never returns, so the value returned from
psci_sys_poweroff() is meaningless, but that function must return
some value to have the right prototype for a notifier callback.

Cover-letter:
firmware/psci: Switch to the sys-off handler API
I want to convert the axp20x PMIC poweroff handler to use the sys-off
API, so it can be used as a fallback for the SBI poweroff handler on
RISC-V. But the PSCI poweroff handler still uses pm_power_off, so done
alone, this conversion would cause the axp20x callback to be called
first, before the PSCI poweroff handler.

In order to prevent this change in behavior, the PSCI poweroff handler
needs to be converted to the sys-off API first, at a higher priority.

This series performs the conversion, after accounting for the fact that
the PSCI poweroff handler is registered quite early during boot.

The first patch is a dependency for both this series and the SBI
series[1], so I would like to get at least patch 1 merged soon.

[1]: https://lore.kernel.org/lkml/20221228161915.13194-1-samuel@sholland.org/
END

Series-version: 2

Series-changes: 2
 - Update commit messages

Series-to: Mark Rutland <mark.rutland@arm.com>
Series-to: Lorenzo Pieralisi <lpieralisi@kernel.org>
Series-to: Sudeep Holla <sudeep.holla@arm.com>
Series-to: Dmitry Osipenko <dmitry.osipenko@collabora.com>
Series-to: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

Signed-off-by: Samuel Holland <samuel@sholland.org>