riscv: sbi: Switch to the sys-off handler API
Any other poweroff handlers registered at the default priority will
run before the legacy pm_power_off() function. Register the SBI
poweroff handler with the correct priority to ensure it runs first.

Cover-letter:
riscv: sbi: Switch to the sys-off handler API
I want to convert the axp20x PMIC poweroff handler to use the sys-off
API, so it can be used as a fallback for if the SBI poweroff handler
is unavailable. But the SBI poweroff handler still uses pm_power_off, so
done alone, this would cause the axp20x callback to be called first,
before the SBI poweroff handler has a chance to run.

In order to prevent this change in behavior, the SBI poweroff handler
needs to be converted to the sys-off API first, at a higher priority.

This series performs the conversion, after accounting for the fact that
the SBI poweroff handler is registered quite early during boot.

The first patch is a dependency for both this series and the PSCI
series[1], so I would like to get at least patch 1 merged soon.

[1]: https://lore.kernel.org/lkml/20221105214841.7828-1-samuel@sholland.org/
END

Series-to: Palmer Dabbelt <palmer@dabbelt.com>
Series-to: Dmitry Osipenko <dmitry.osipenko@collabora.com>
Series-to: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

Signed-off-by: Samuel Holland <samuel@sholland.org>